created separate file for cctx query with rate limit
