+++ x/crosschain/keeper/grpc_query_cctx_test.go
@@ -185,36 +185,6 @@ func TestKeeper_CctxListPending(t *testing.T) {
 	})
 }
 
-func TestKeeper_CctxListPendingWithRateLimit(t *testing.T) {
-	t.Run("should fail for empty req", func(t *testing.T) {
-		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
-		_, err := k.CctxListPendingWithinRateLimit(ctx, nil)
-		require.ErrorContains(t, err, "invalid request")
-	})
-
-	t.Run("should fail if limit is too high", func(t *testing.T) {
-		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
-		_, err := k.CctxListPendingWithinRateLimit(ctx, &types.QueryListCctxPendingWithRateLimitRequest{Limit: keeper.MaxPendingCctxs + 1})
-		require.ErrorContains(t, err, "limit exceeds max limit of")
-	})
-
-	t.Run("should fail if no TSS", func(t *testing.T) {
-		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
-		_, err := k.CctxListPendingWithinRateLimit(ctx, &types.QueryListCctxPendingWithRateLimitRequest{Limit: 1})
-		require.ErrorContains(t, err, "tss not found")
-	})
-
-	t.Run("should return empty list if no nonces", func(t *testing.T) {
-		k, ctx, _, zk := keepertest.CrosschainKeeper(t)
-
-		//  set TSS
-		zk.ObserverKeeper.SetTSS(ctx, sample.Tss())
-
-		_, err := k.CctxListPendingWithinRateLimit(ctx, &types.QueryListCctxPendingWithRateLimitRequest{Limit: 1})
-		require.ErrorContains(t, err, "pending nonces not found")
-	})
-}
-
 func TestKeeper_ZetaAccounting(t *testing.T) {
 	t.Run("should error if zeta accounting not found", func(t *testing.T) {
 		k, ctx, _, _ := keepertest.CrosschainKeeper(t)
